fix: do not append to configured log message (#111)
There was a recent bug introduced recently that caused the `with error:`
section of the message to be appended every time after a request with an
error happened. This is because it was appending to the configuration
message and not to a copy of it.

This adds a test and fixes the issue by creating a copy before
appending and using that to log.